put this here; non-functional experiment
I wanted my enums to be c-style so that they were just
really the underlying number that CoAP uses, but I couldn't
figure out a way to make the unassigned (but potentially
used by other that this library will interact with in the
future) work in a way that was totally safe. Maybe I'll
get back to this idea eventually.